Hydro example tests : use "H. LEV" instead of index 26
